Fix compiler warning

+++ usermods/PIR_sensor_switch/usermod_PIR_sensor_switch.h
@@ -533,7 +533,7 @@ bool PIRsensorSwitch::readFromConfig(JsonObject &root)
 
   JsonArray pins = top["pin"];
   if (!pins.isNull()) {
-    for (int i = 0; i < PIR_SENSOR_MAX_SENSORS; i++)
+    for (size_t i = 0; i < PIR_SENSOR_MAX_SENSORS; i++)
       if (i < pins.size()) PIRsensorPin[i] = pins[i] | PIRsensorPin[i];
   } else {
     PIRsensorPin[0] = top["pin"] | oldPin[0];
